src/test/java/uk/gov/hmcts/pdm/publicdisplay/manager/web/users/UserAddValidatorTest.java
/*
 * Copyrights and Licenses
 * 
 * Copyright (c) 2015-2016 by the Ministry of Justice. All rights reserved. Redistribution and use
 * in source and binary forms, with or without modification, are permitted provided that the
 * following conditions are met: - Redistributions of source code must retain the above copyright
 * notice, this list of conditions and the following disclaimer. - Redistributions in binary form
 * must reproduce the above copyright notice, this list of conditions and the following disclaimer
 * in the documentation and/or other materials provided with the distribution. - Products derived
 * from this software may not be called "XHIBIT Public Display Manager" nor may
 * "XHIBIT Public Display Manager" appear in their names without prior written permission of the
 * Ministry of Justice. - Redistributions of any form whatsoever must retain the following
 * acknowledgment: "This product includes XHIBIT Public Display Manager." This software is provided
 * "as is" and any expressed or implied warranties, including, but not limited to, the implied
 * warranties of merchantability and fitness for a particular purpose are disclaimed. In no event
 * shall the Ministry of Justice or its contributors be liable for any direct, indirect, incidental,
 * special, exemplary, or consequential damages (including, but not limited to, procurement of
 * substitute goods or services; loss of use, data, or profits; or business interruption). However
 * caused any on any theory of liability, whether in contract, strict liability, or tort (including
 * negligence or otherwise) arising in any way out of the use of this software, even if advised of
 * the possibility of such damage.
 */

package uk.gov.hmcts.pdm.publicdisplay.manager.web.users;

import org.easymock.Capture;
import org.easymock.EasyMockExtension;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.context.MessageSource;
import org.springframework.context.support.ReloadableResourceBundleMessageSource;
import org.springframework.test.util.ReflectionTestUtils;
import org.springframework.validation.BeanPropertyBindingResult;
import org.springframework.validation.BindingResult;
import uk.gov.hmcts.pdm.publicdisplay.common.test.AbstractJUnit;
import uk.gov.hmcts.pdm.publicdisplay.manager.security.UserRole;
import uk.gov.hmcts.pdm.publicdisplay.manager.service.UserDetailsService;

import java.util.Locale;

import static org.easymock.EasyMock.capture;
import static org.easymock.EasyMock.createMock;
import static org.easymock.EasyMock.expect;
import static org.easymock.EasyMock.newCapture;
import static org.easymock.EasyMock.replay;
import static org.easymock.EasyMock.verify;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertNotNull;

/**
 * The Class UserAddValidatorTest.
 * 
 * @author harrism
 *
 */
@ExtendWith(EasyMockExtension.class)
class UserAddValidatorTest extends AbstractJUnit {

    private static final String NOT_EQUAL = "Not equal";

    private static final String NULL = "Null";

    private static final String TRUE = "True";

    /** The class under test. */
    private UserAddValidator classUnderTest;

    /** The mock user details service. */
    private UserDetailsService mockUserDetailsService;

    /**
     * Setup.
     */
    @BeforeEach
    public void setup() {
        // Create a new version of the class under test
        classUnderTest = new UserAddValidator();

        // Setup the mock version of the called classes
        mockUserDetailsService = createMock(UserDetailsService.class);
        MessageSource mockMessageSource = new ReloadableResourceBundleMessageSource() {
            @Override
            public String getMessageInternal(String code, Object[] args, Locale locale) {
                return "Test";
            }
        };

        // Map the mock to the class under tests called class
        ReflectionTestUtils.setField(classUnderTest, "userDetailsService", mockUserDetailsService);
        ReflectionTestUtils.setField(classUnderTest, "messageSource", mockMessageSource);
    }

    /**
     * Test validate success.
     */
    @Test
    void testValidateSuccess() {
        // Capture the username
        final Capture<String> capturedUserName = newCapture();

        final UserAddCommand userCommand = getUserCommand();
        final BindingResult errors = new BeanPropertyBindingResult(userCommand, "userAddCommand");

        // Define a mock version of the called methods
        expect(mockUserDetailsService.isUserWithUserName(capture(capturedUserName)))
            .andReturn(false);
        replay(mockUserDetailsService);

        // Perform the test
        classUnderTest.validate(userCommand, errors);

        // Check the results
        assertNotNull(errors, NULL);
        assertFalse(errors.hasErrors(), TRUE);
        assertEquals(userCommand.getUserName(), capturedUserName.getValue(), NOT_EQUAL);

        // Verify the mocks used in this method were called
        verify(mockUserDetailsService);
    }

    /**
     * Test validate user exists failure.
     */
    @Test
    void testValidateUserExistsFailure() {
        // Capture the username
        final Capture<String> capturedUserName = newCapture();

        final UserAddCommand userCommand = getUserCommand();
        final BindingResult errors = new BeanPropertyBindingResult(userCommand, "userAddCommand");

        // Define a mock version of the called methods
        expect(mockUserDetailsService.isUserWithUserName(capture(capturedUserName)))
            .andReturn(true);
        replay(mockUserDetailsService);

        // Perform the test
        classUnderTest.validate(userCommand, errors);

        // Check the results
        assertNotNull(errors, NULL);
        assertEquals(1, errors.getErrorCount(), NOT_EQUAL);
        assertEquals(userCommand.getUserName(), capturedUserName.getValue(), NOT_EQUAL);

        // Verify the mocks used in this method were called
        verify(mockUserDetailsService);
    }

    /**
     * Gets the test user command.
     *
     * @return the test user command
     */
    private UserAddCommand getUserCommand() {
        final UserAddCommand usercommand = new UserAddCommand();
        usercommand.setUserName("USERNAME");
        usercommand.setUserRole(UserRole.ROLE_ADMIN);
        return usercommand;
    }
}
